pyproject.toml
[project]
name = "medseg-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "dvc>=3.66.1",
    "evidently>=0.7.20",
    "fastapi>=0.135.1",
    "gdown>=5.2.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "segmentation-models-pytorch>=0.5.0",
    "streamlit>=1.55.0",
    "tensorboard>=2.20.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
# Pins these packages to the custom index defined below
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true  # Prevents other packages from accidentally using this index
